renames products modules to CSR layers
