+++ api/specs/web-server/_products.py
@@ -18,7 +18,7 @@ from models_library.api_schemas_webserver.product import (
 )
 from models_library.generics import Envelope
 from simcore_service_webserver._meta import API_VTAG
-from simcore_service_webserver.products._handlers import (
+from simcore_service_webserver.products._rest import (
     _ProductsRequestParams,
     _ProductTemplateParams,
 )
