+++ services/web/server/src/simcore_service_webserver/products/_repository.py
@@ -16,7 +16,7 @@ from simcore_postgres_database.utils_products_prices import (
 
 from ..db.base_repository import BaseRepository
 from ..db.models import products
-from ._model import Product
+from ._models import Product
 
 _logger = logging.getLogger(__name__)
