+++ services/web/server/src/simcore_service_webserver/products/_rpc.py
@@ -6,7 +6,7 @@ from models_library.products import CreditResultGet, ProductName
 from servicelib.rabbitmq import RPCRouter
 
 from ..rabbitmq import get_rabbitmq_rpc_server
-from . import _api
+from . import _service
 
 router = RPCRouter()
 
@@ -18,7 +18,7 @@ async def get_credit_amount(
     dollar_amount: Decimal,
     product_name: ProductName,
 ) -> CreditResultGet:
-    credit_result_get: CreditResultGet = await _api.get_credit_amount(
+    credit_result_get: CreditResultGet = await _service.get_credit_amount(
         app, dollar_amount=dollar_amount, product_name=product_name
     )
     return credit_result_get
